More RampackMatcher prepwork
